trabalho2/requirements.txt
scikit-image==0.24.0
numpy==2.2.4
pillow==11.1.0
opencv-python==4.11.0.86
matplotlib==3.9.4